requirements.txt
Django==2.0.6
django-widgets-improved==1.5.0
djangorestframework==3.8.2
django-object-actions==0.10.0
django-jet==1.0.7
django-filter==1.1.0
django-oauth-toolkit==1.2.0
drf-extensions==0.3.1
python-decouple==3.1
django_countries==5.3.1
django-phonenumber-field==2.0.0
Pillow==5.1.0
django-guardian==1.4.9
